+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Admin settings: show and change all app settings via occ
 - Admins can delete the stored codes of one or all users via occ
+- Notify the user if an admin (de)activated this 2FA provider or the primary email address was deleted
 
 ### Fixed